requirements.txt
flake8==3.8.3
pre-commit==1.15.1
dbus-python==1.2.16
dbus-next==0.2.3
twine
wheel
